refactor: enums
