+++ frontend/src/enums/boards/actions.ts
@@ -1,0 +1,14 @@
+export enum BoardAction {
+  UPDATECARDPOSITION,
+  VOTE,
+  UNMERGE,
+  MERGE,
+  ADDCARD,
+  UPDATECARD,
+  DELETECARD,
+  ADDCOMMENT,
+  DELETECOMMENT,
+  UPDATECOMMENT,
+  UPDATEBOARDUSERS,
+  UPDATEPHASE,
+}
