+++ frontend/src/enums/timer/status.ts
