+++ frontend/src/enums/votes/actions.ts
@@ -1,0 +1,4 @@
+export enum Action {
+  Add = 'add',
+  Remove = 'remove',
+}
