+++ frontend/src/hooks/useSocketIO.ts
@@ -6,6 +6,7 @@ import { io, Socket } from 'socket.io-client';
 
 import { BOARD_PHASE_SERVER_SENT, NEXT_PUBLIC_BACKEND_URL } from '@/constants';
 import { ROUTES } from '@/constants/routes';
+import { BoardAction } from '@/enums/boards/actions';
 import { ToastStateEnum } from '@/enums/toasts/toast-types';
 import useBoard from '@/hooks/useBoard';
 import useCards from '@/hooks/useCards';
@@ -29,21 +30,6 @@ import EventCallback from '@/types/events/event-callback.type';
 import { ListenEvent } from '@/types/events/listen-event.type';
 import VoteDto from '@/types/vote/vote.dto';
 import isEmpty from '@/utils/isEmpty';
-
-enum BoardAction {
-  UPDATECARDPOSITION,
-  VOTE,
-  UNMERGE,
-  MERGE,
-  ADDCARD,
-  UPDATECARD,
-  DELETECARD,
-  ADDCOMMENT,
-  DELETECOMMENT,
-  UPDATECOMMENT,
-  UPDATEBOARDUSERS,
-  UPDATEPHASE,
-}
 
 interface SocketInterface {
   socketId?: string;
