+++ frontend/src/types/timer/timer-state.dto.ts
@@ -1,5 +1,5 @@
+import { TimerStatus } from '@/enums/timer/status';
 import { TimeDto } from '@/types/timer/time.dto';
-import { TimerStatus } from '@/types/timer/timer-status.enum';
 
 export interface TimerStateDto {
   status: TimerStatus;
